feat: implement connection and segmentation
