+++ src/types/network-modules/tcpModule.ts
@@ -61,14 +61,7 @@ export class TcpModule {
     const flags = new Flags().withSyn();
     const srcPort: Port = this.getNextPortNumber();
     const seqNum = getInitialSeqNumber();
-    const synSegment = new TcpSegment(
-      srcPort,
-      dstPort,
-      seqNum,
-      0,
-      flags,
-      new Uint8Array(),
-    );
+    const synSegment = new TcpSegment(srcPort, dstPort, seqNum, 0, flags);
     // Send SYN
     sendIpPacket(this.host, dstHost, synSegment);
 
@@ -299,227 +292,6 @@ export class TcpListener {
     const queue = this.tcpModule.initNewQueue(this.port, ipAndPort);
 
     return new TcpSocket(this.host, this.port, dst, segment.sourcePort, queue);
-  }
-}
-
-// TODO: add overflow checks
-class TcpState {
-  private srcPort: Port;
-  private dstPort: Port;
-
-  // Buffer of data received
-  private readBuffer = new BytesBuffer(MAX_BUFFER_SIZE);
-  private readClosed = false;
-
-  // Buffer of data to be sent
-  private writeBuffer = new BytesBuffer(MAX_BUFFER_SIZE);
-  private writeClosed = false;
-
-  // See https://datatracker.ietf.org/doc/html/rfc9293#section-3.3.1
-  // SND.UNA
-  private sendUnacknowledged: number;
-  // SND.NXT
-  private sendNext: number;
-  // SND.WND
-  private sendWindow: number;
-  // SND.UP
-  // private sendUrgentPointer;
-  // SND.WL1
-  private seqNumForLastWindowUpdate: number;
-  // SND.WL2
-  private ackNumForLastWindowUpdate: number;
-  // ISS
-  private initialSendSeqNum;
-
-  // RCV.NXT
-  private recvNext: number;
-  // RCV.WND
-  private recvWindow: number;
-
-  // IRS
-  private initialRecvSeqNum: number;
-
-  constructor(srcPort: Port, dstPort: Port) {
-    this.srcPort = srcPort;
-    this.dstPort = dstPort;
-    this.initialSendSeqNum = getInitialSeqNumber();
-  }
-
-  processSynAck(segment: TcpSegment) {
-    if (!segment.flags.syn || !segment.flags.ack) {
-      return false;
-    }
-    if (segment.acknowledgementNumber !== this.initialSendSeqNum + 1) {
-      return false;
-    }
-  }
-
-  read(output: Uint8Array): number {
-    const readLength = this.readBuffer.read(output);
-    if (readLength === 0 && this.readClosed) {
-      return -1;
-    }
-    return readLength;
-  }
-
-  write(input: Uint8Array): number {
-    if (this.writeClosed) {
-      throw new Error("write closed");
-    }
-    return this.writeBuffer.write(input);
-  }
-
-  recvSegment(receivedSegment: TcpSegment) {
-    // Sanity check: ports match with expected
-    if (
-      receivedSegment.destinationPort !== this.dstPort ||
-      receivedSegment.sourcePort !== this.srcPort
-    ) {
-      throw new Error("segment not for this socket");
-    }
-    // Check the sequence number is valid
-    const segSeq = receivedSegment.sequenceNumber;
-    const segLen = receivedSegment.data.length;
-    if (!this.isSeqNumValid(segSeq, segLen)) {
-      return false;
-    }
-
-    // TODO: check RST flag
-
-    // For now this only accepts ACK segments.
-    // 3WHS are handled outside of this function
-    if (!receivedSegment.flags.ack) {
-      return false;
-    }
-    if (!this.recvAck(receivedSegment)) {
-      return false;
-    }
-
-    // Process the segment data
-    // NOTE: for simplicity, we ignore cases where RCV.NXT != SEG.SEQ
-    if (receivedSegment.sequenceNumber === this.recvNext) {
-      let receivedData = receivedSegment.data;
-      // NOTE: for simplicity, we ignore cases where the data is only partially
-      // inside the window
-      if (receivedData.length > this.recvWindow) {
-        throw new Error("buffer overflow");
-      }
-      this.readBuffer.write(receivedData);
-      this.recvNext = receivedSegment.sequenceNumber + receivedData.length;
-      this.recvWindow = MAX_BUFFER_SIZE - this.readBuffer.bytesAvailable();
-    }
-
-    // If FIN, mark read end as closed
-    if (receivedSegment.flags.fin) {
-      this.readClosed = true;
-    }
-    return true;
-  }
-
-  nextSegment(): TcpSegment | null {
-    const segment = new TcpSegment(
-      0,
-      0,
-      this.seqNum,
-      this.ackNum,
-      new Flags(),
-      new Uint8Array(),
-    );
-    return segment;
-  }
-
-  // utils
-
-  private isSeqNumValid(segSeq: number, segLen: number) {
-    const lengthIsZero = segLen === 0;
-    const windowIsZero = this.recvWindow === 0;
-
-    if (lengthIsZero && windowIsZero) {
-      return segSeq === this.recvNext;
-    } else if (lengthIsZero && !windowIsZero) {
-      return this.isInReceiveWindow(segSeq);
-    } else if (!lengthIsZero && windowIsZero) {
-      return false;
-    } else {
-      return (
-        this.isInReceiveWindow(segSeq) ||
-        this.isInReceiveWindow(segSeq + segLen - 1)
-      );
-    }
-  }
-
-  private isInReceiveWindow(n: number) {
-    return this.recvNext <= n && n < this.recvNext + this.recvWindow;
-  }
-
-  private recvAck(segment: TcpSegment) {
-    // From https://datatracker.ietf.org/doc/html/rfc9293#section-3.10.7.4-2.5.2.2.2.3.1
-    // If the ACK is for a packet not yet sent, drop it
-    if (this.sendNext < segment.acknowledgementNumber) {
-      return false;
-    }
-    // If SND.UNA < SEG.ACK =< SND.NXT, set SND.UNA <- SEG.ACK
-    if (this.sendUnacknowledged <= segment.acknowledgementNumber) {
-      this.sendUnacknowledged = segment.acknowledgementNumber;
-      if (this.isSegmentNewer(segment)) {
-        // set SND.WND <- SEG.WND, set SND.WL1 <- SEG.SEQ, and set SND.WL2 <- SEG.ACK.
-        this.sendWindow = segment.window;
-        this.seqNumForLastWindowUpdate = segment.sequenceNumber;
-        this.ackNumForLastWindowUpdate = segment.acknowledgementNumber;
-      }
-    }
-  }
-
-  private isSegmentNewer(segment: TcpSegment): boolean {
-    // Since both SEQ and ACK numbers are monotonic, we can use
-    // them to determine if the segment is newer than the last
-    // one that was used for updating the window
-    //
-    // SND.WL1 < SEG.SEQ or (SND.WL1 = SEG.SEQ and SND.WL2 =< SEG.ACK)
-    return (
-      this.seqNumForLastWindowUpdate < segment.sequenceNumber ||
-      (this.seqNumForLastWindowUpdate === segment.sequenceNumber &&
-        this.ackNumForLastWindowUpdate <= segment.acknowledgementNumber)
-    );
-  }
-}
-
-class BytesBuffer {
-  private buffer: Uint8Array;
-  private length: number;
-
-  constructor(size: number) {
-    this.buffer = new Uint8Array(size);
-    this.length = 0;
-  }
-
-  read(output: Uint8Array): number {
-    const readLength = Math.min(this.length, output.length);
-    if (readLength == 0) {
-      return 0;
-    }
-    output.set(this.buffer.subarray(0, readLength));
-    this.buffer.copyWithin(0, readLength, this.length);
-    this.length -= readLength;
-    return readLength;
-  }
-
-  write(data: Uint8Array): number {
-    const newLength = this.length + data.length;
-    if (newLength > this.buffer.length) {
-      return 0;
-    }
-    this.buffer.set(data, this.length);
-    this.length = newLength;
-    return data.length;
-  }
-
-  bytesAvailable() {
-    return this.length;
-  }
-
-  isEmpty() {
-    return this.bytesAvailable() === 0;
   }
 }
